+++ src/main/java/com/ing/data/cassandra/jdbc/utils/ErrorConstants.java
@@ -416,6 +416,20 @@ public final class ErrorConstants {
      */
     public static final String CANNOT_OPEN_SOURCE_FILE = "Could not open '%s': %s";
 
+    /**
+     * Error message used when the execution of the special CQL command {@code COPY TO} failed while writing the CSV
+     * file. This message is a template expecting the name of the file and the error message (example:
+     * {@code String.format(CANNOT_WRITE_CSV_FILE, "target.csv", "access refused")}).
+     */
+    public static final String CANNOT_WRITE_CSV_FILE = "Could not write CSV file '%s': %s";
+
+    /**
+     * Error message used when the special CQL command {@code COPY} contains at least one unknown or unsupported option.
+     * This message is a template expecting the unknown option (example:
+     * {@code String.format(UNSUPPORTED_COPY_OPTIONS, "[OPTION1, OPTION2]")}).
+     */
+    public static final String UNSUPPORTED_COPY_OPTIONS = "Command COPY used with unknown or unsupported options: %s";
+
     private ErrorConstants() {
         // Private constructor to hide the public one.
     }